+++ src/test/java/faang/school/accountservice/service/BalanceServiceTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -37,8 +38,8 @@ public class BalanceServiceTest {
     void setUp() {
         balance = new Balance();
         balance.setId(UUID.fromString("2a5c971f-69c6-4dc2-a341-cd322ab3da25"));
-        balance.setAuthorization(0.0);
-        balance.setActual(0.0);
+        balance.setAuthorization(BigDecimal.valueOf(0));
+        balance.setActual(BigDecimal.valueOf(0));
 
         account = new Account();
         account.setId(UUID.fromString("7e91477c-0121-4123-a112-6c98a4413e1b"));
@@ -49,8 +50,8 @@ public class BalanceServiceTest {
     @DisplayName("Balance service: create balance")
     public void testCreateBalance_checkExecute() {
         when(balanceService.createBalance()).thenReturn(balance);
-        assertEquals(0.0, balance.getAuthorization());
-        assertEquals(0.0, balance.getActual());
+        assertEquals(BigDecimal.valueOf(0), balance.getAuthorization());
+        assertEquals(BigDecimal.valueOf(0), balance.getActual());
 
         balanceService.createBalance();
         verify(balanceRepository).save(any(Balance.class));
